ae_mind/requirements.txt
# Core Dependencies
aiohttp>=3.8.0
cryptography>=3.4.7
PyJWT>=2.3.0
psutil>=5.8.0
pytest>=6.2.5
pytest-asyncio>=0.16.0
aiofiles>=0.8.0

# Security
bcrypt>=3.2.0
python-jose>=3.3.0

# Monitoring
prometheus-client>=0.12.0
statsd>=3.3.0

# Testing
pytest-cov>=2.12.0
pytest-mock>=3.6.1
asynctest>=0.13.0

# Documentation
Sphinx>=4.3.0
sphinx-rtd-theme>=1.0.0

# Parser Dependencies
spacy
beautifulsoup4
pydantic
tabulate
PyYAML
pandas
pyarrow
transformers
torch
